requirements.txt
fastapi==0.78.0
minio==7.1.9
Pillow==9.1.1
requests==2.28.0

SQLAlchemy~=1.4.39
starlette~=0.19.1
pydantic~=1.9.1